Moved existing RefreshEndpoint logic to the new RefreshSupport class. Fixes #101.
